requirements.txt
pycryptodome==3.9.8
PyQt5==5.15.1
PyQt5-sip==12.8.1
